fix findbugs warning 'No relationship between generic parameter and method argument'

+++ kogito-jbpm/jbpm-persistence-jpa/src/main/java/org/jbpm/persistence/timer/GlobalJPATimerJobFactoryManager.java
@@ -120,7 +120,7 @@ public class GlobalJPATimerJobFactoryManager implements TimerJobFactoryManager {
         return singleTimerInstances.values();
     }
     
-    public Collection<TimerJobInstance> getTimerJobInstances(Integer sessionId) {
+    public Collection<TimerJobInstance> getTimerJobInstances(Long sessionId) {
         Map<Long, TimerJobInstance> sessionTimerJobs = timerInstances.get(sessionId);
         if (sessionTimerJobs == null) {
             return Collections.EMPTY_LIST;
